Rename ProductCard to ProductGridItem
