+++ apps/store/src/blocks/ProductGridBlock.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import { storyblokEditable } from '@storyblok/react'
-import type { ProductCardBlockProps } from '@/blocks/ProductCardBlock'
-import { ProductCardBlock } from '@/blocks/ProductCardBlock'
+import type { ProductCardBlockProps } from '@/blocks/ProductGridItemBlock/ProductGridItemBlock'
+import { ProductGridItemBlock } from '@/blocks/ProductGridItemBlock/ProductGridItemBlock'
 import { ProductGrid } from '@/components/ProductGrid/ProductGrid'
 import type { ExpectedBlockType, SbBaseBlockProps } from '@/services/storyblok/storyblok'
 
@@ -15,7 +15,7 @@ export const ProductGridBlock = ({ blok }: ProductGridBlockProps) => {
   return (
     <ProductGrid title={blok.title} {...storyblokEditable(blok)}>
       {blok.items.map((nestedBlock) => (
-        <ProductCardBlock key={nestedBlock._uid || nestedBlock.title} blok={nestedBlock} />
+        <ProductGridItemBlock key={nestedBlock._uid || nestedBlock.title} blok={nestedBlock} />
       ))}
     </ProductGrid>
   )
