+++ apps/store/src/blocks/ProductGridItemBlock/ProductGridItemBlock.tsx
@@ -1,10 +1,11 @@
 'use client'
+
 import { storyblokEditable } from '@storyblok/react'
 import { useProductMetadata } from '@/components/LayoutWithMenu/productMetadataHooks'
-import { type LinkType, ProductCard } from '@/components/ProductCard/ProductCard'
+import { ProductGridItem } from '@/components/ProductGridItem/ProductGridItem'
 import type { LinkField, SbBaseBlockProps, StoryblokAsset } from '@/services/storyblok/storyblok'
 import { getImgSrc, getLinkFieldURL } from '@/services/storyblok/Storyblok.helpers'
-import { isSameLink } from '@/utils/url'
+import { getLinkType } from './ProductGridItemBlock.helpers'
 
 export type ImageSize = {
   aspectRatio?: '1 / 1' | '3 / 2' | '4 / 3' | '5 / 4' | '2 / 3' | '3 / 4' | '4 / 5'
@@ -20,20 +21,21 @@ export type ProductCardBlockProps = SbBaseBlockProps<
   } & ImageSize
 >
 
-export const ProductCardBlock = ({ blok }: ProductCardBlockProps) => {
+export const ProductGridItemBlock = ({ blok }: ProductCardBlockProps) => {
   const link = getLinkFieldURL(blok.link, blok.title)
   const productMetadata = useProductMetadata()
   const linkType = getLinkType(productMetadata, link)
 
   if (linkType === 'content') {
     console.warn(
-      '[ProductCardBlock]: provided "link" does not refer to a product neither a category. Skipping ProductCard render!',
+      '[ProductGridItemBlock]: provided "link" does not refer to a product neither a category. Skipping ProductGridItem render!',
     )
+
     return null
   }
 
   return (
-    <ProductCard
+    <ProductGridItem
       title={blok.title}
       subtitle={blok.subtitle}
       image={{ src: getImgSrc(blok.image.filename), alt: blok.image.alt, priority: blok.priority }}
@@ -42,23 +44,4 @@ export const ProductCardBlock = ({ blok }: ProductCardBlockProps) => {
       {...storyblokEditable(blok)}
     />
   )
-}
-
-const getLinkType = (
-  productMetadata: ReturnType<typeof useProductMetadata> = [],
-  link: string,
-): LinkType | 'content' => {
-  const isProductLink = productMetadata?.some((product) => isSameLink(link, product.pageLink))
-  if (isProductLink) {
-    return 'product'
-  }
-
-  const isCategoryLink = productMetadata?.some(
-    (product) => product.categoryPageLink && isSameLink(link, product.categoryPageLink),
-  )
-  if (isCategoryLink) {
-    return 'category'
-  }
-
-  return 'content'
 }
